Improve post-acceptance error checking

+++ post-acceptance.sh
@@ -6,6 +6,13 @@
 VERSION=$(grep '^Version:' DESCRIPTION | sed -e 's/Version: *//')
 CRANCOMMIT=$(cat CRAN-SUBMISSION)
 
+if [[ -z $VERSION || -z $CRANCOMMIT ]]; then
+  echo "VERSION or CRANCOMMIT was empty"
+  echo "VERSION: '$VERSION'"
+  echo "CRANCOMMIT: '$CRANCOMMIT'"
+  exit 1
+fi
+
 git tag -m "Version $VERSION on CRAN" -a v$VERSION $CRANCOMMIT
 git push --tags
 rm CRAN-SUBMISSION
@@ -13,3 +20,4 @@ rm CRAN-SUBMISSION
 Rscript -e 'revdepcheck::revdep_reset()'
 Rscript -e 'usethis::use_github_release()'
 
+exit 0
